requirements.txt
streamlit
langchain
langchain-groq
whisper
torch
transformers
sentence-transformers
faiss-cpu
fpdf
python-dotenv